May 31, 2026, 2:01 PM

+++ set.h
@@ -88,6 +88,6 @@ public:
 
 private:
   T* m_values;
-  int m_size;
+  size_t m_size;
   size_t m_capacity;
 };
